Modifies the burn process to leave thin dead moss layer.
This is less correct, but the ground is not recovering after fire.
e.g. there is litterfall following fire, but never any recreation
of shallow organic layers.

We linked living moss soil layer with the living moss PFT carbon pool and
forced the dead moss layer to not burn completely so it can recover
after fire.

We attempted to leave one thin layer of each type post fire, but this
was only partially successful.