Parallel I/O functional for two variables
This is a placeholder commit for proof of concept for parallel I/O.
hdf5-1.8.19, netcdf-4.4.1.1, both manually built for parallel.
One of the oddities of this version of hdf5 is that it installs
to a directory named hdf5-1.10.1. There are some complexities with
the installs, to be explained further on the wiki.