Modifies NC output file creation to use specified data types
Prior to this, all files were created with the type NC_DOUBLE.
When the put_vara_type() functions were used, this was fine. Now
that the model uses the general put_vara() function, this caused
issues with non-double variables.

Having a manual tag in the output_spec file is not very robust,
so this approach may need to be replaced in the future.